sweeper for relay
