+++ internal/service/mailmanager/archive.go
@@ -22,7 +22,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs/fwdiag"
 	"github.com/hashicorp/terraform-provider-aws/internal/framework"
@@ -30,8 +29,6 @@ import (
 	fwtypes "github.com/hashicorp/terraform-provider-aws/internal/framework/types"
 	"github.com/hashicorp/terraform-provider-aws/internal/retry"
 	"github.com/hashicorp/terraform-provider-aws/internal/smerr"
-	"github.com/hashicorp/terraform-provider-aws/internal/sweep"
-	sweepfw "github.com/hashicorp/terraform-provider-aws/internal/sweep/framework"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
 	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
 	"github.com/hashicorp/terraform-provider-aws/names"
@@ -318,26 +315,4 @@ func (m *archiveRetentionModel) Flatten(ctx context.Context, v any) diag.Diagnos
 			fmt.Sprintf("archive retention flatten: unexpected type %T", v))
 	}
 	return diags
-}
-
-func sweepArchives(ctx context.Context, client *conns.AWSClient) ([]sweep.Sweepable, error) {
-	input := mailmanager.ListArchivesInput{}
-	conn := client.MailManagerClient(ctx)
-	var sweepResources []sweep.Sweepable
-
-	pages := mailmanager.NewListArchivesPaginator(conn, &input)
-	for pages.HasMorePages() {
-		page, err := pages.NextPage(ctx)
-		if err != nil {
-			return nil, smarterr.NewError(err)
-		}
-
-		for _, v := range page.Archives {
-			sweepResources = append(sweepResources, sweepfw.NewSweepResource(newArchiveResource, client,
-				sweepfw.NewAttribute(names.AttrID, aws.ToString(v.ArchiveId))),
-			)
-		}
-	}
-
-	return sweepResources, nil
 }
